Add email verify get API to dashboard recipe
