+++ recipe/dashboard/constants.go
@@ -5,3 +5,4 @@ const validateKeyAPI = "/api/key/validate"
 const usersListGetAPI = "/api/users"
 const usersCountAPI = "/api/users/count"
 const userAPI = "/api/user"
+const userEmailVerifyAPI = "/api/user/email/verify"
